Refactor from feedback
