+++ html/index.html
@@ -524,6 +524,7 @@
     <script src="js/routes/job.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/settings.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/grid.js?v=VERSION_PLACEHOLDER"></script>
+    <script src="js/routes/grid_health.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/grid_metrics.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/gridmembers.js?v=VERSION_PLACEHOLDER"></script>
     <script src="js/routes/terms.js?v=VERSION_PLACEHOLDER"></script>
