+++ html/js/i18n.js
@@ -738,6 +738,21 @@ const i18n = {
       esHealthNoIssues: 'No issues detected. All Elasticsearch health indicators are green.',
       esHealthPrimary: 'primary',
       esHealthReplica: 'replica',
+      esHealthReportCanAllocate: 'Can allocate',
+      esHealthReportGenerated: 'Generated',
+      esHealthReportNoSample: 'No allocation explanation sampled for {value}',
+      esHealthReportNode: 'Node',
+      esHealthReportNone: '(none)',
+      esHealthReportSampledShard: 'Sampled shard for {value}:',
+      esHealthReportSectionIndicators: '--- Indicators (_health_report) ---',
+      esHealthReportSectionNodes: '--- Nodes (_cat/nodes) ---',
+      esHealthReportSectionSettings: '--- Cluster Settings (_cluster/settings) ---',
+      esHealthReportSectionShards: '--- Unassigned Shards (_cat/shards, _cluster/allocation/explain) ---',
+      esHealthReportShard: 'Shard',
+      esHealthReportSince: ', since {value}',
+      esHealthReportTitle: 'ELASTICSEARCH HEALTH REPORT',
+      esHealthReportTotals: 'Total: {total} unassigned ({primaries} primary, {replicas} replica)',
+      esHealthReportUnassignedReason: 'Unassigned reason',
       esHealthShow: 'View Elasticsearch cluster health details',
       esHealthTitle: 'Elasticsearch Health',
       esHealthUnavailable: 'Elasticsearch health information is currently unavailable.',
